pyproject.toml
[project]
name = "search-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.5",
    "djangorestframework>=3.15.2",
    "faker>=35.1.0",
    "meilisearch>=0.33.1",
]
